+++ src/main/java/com/bda/TrabajoPractico/servicios/interfaces/AlquilerService.java
@@ -2,5 +2,5 @@ package com.bda.TrabajoPractico.servicios.interfaces;
 
 import com.bda.TrabajoPractico.Entidades.Alquiler;
 
-public interface AlquilerService extends Service<Alquiler,Integer> {
+public interface AlquilerService extends Service<Alquiler,Long> {
 }
